project0037-dimensionality-reduction-techniques/requirements.txt
pandas
matplotlib
scikit-learn
umap-learn
